+++ dev/App.tsx
@@ -1,7 +1,7 @@
-import { createSignal, type Component, createResource, Show } from 'solid-js';
+import { createSignal, type Component, createResource, Show, createEffect } from 'solid-js';
 
 // query builder lib import
-import { QueryBuilder, type Not_Selection } from 'src';
+import { QueryBuilder, type Not_Selection, type Query } from 'src';
 
 import { ThemeToggle } from './components/ThemeToggle';
 
@@ -21,6 +21,7 @@ const fetchPackageVersion = () => {
 
 const App: Component = () => {
   const [packageVersion] = createResource(fetchPackageVersion);
+  const [qbData, setQbData] = createSignal<Query>(MOCK_QUERY_DATA);
 
   const [disableQB, setDisableQB] = createSignal(false);
   const [showBranches, setShowBranches] = createSignal(true);
@@ -167,7 +168,19 @@ const App: Component = () => {
               addSingleRuleToGroup={addSingleRuleToGroup()}
               showNotToggle={showNotToggle()}
               showBranches={showBranches()}
+              onQueryChangeHandler={query => {
+                setQbData({ ...query });
+                console.log('Query', query);
+              }}
             />
+          </section>
+          <section>
+            <details>
+              <summary>Query Builder JSON Output (show more...)</summary>
+              <pre class="pre">
+                <code>{JSON.stringify(qbData(), null, 2)}</code>
+              </pre>
+            </details>
           </section>
         </section>
         <hr />
@@ -284,7 +297,7 @@ const App: Component = () => {
           <section id="JSON-Data">
             <details>
               <summary>
-                <h5>JSON Data:</h5>
+                <h5>Mock JSON Query Data:</h5>
               </summary>
               <pre class="pre">
                 <code>{JSON.stringify(MOCK_QUERY_DATA, null, 2)}</code>
@@ -313,7 +326,16 @@ const App: Component = () => {
                 <td>
                   Initial query object to be displayed in the Query Builder. The query object should
                   follow the <a href="#JSON-Data">Query object structure</a>. This prop is optional
-                  unless you want to display some initial query.
+                  unless you want to display some initial query. If not provided, it would have some
+                  default query with empty rule.
+                </td>
+              </tr>
+              <tr>
+                <td>onQueryChangeHandler</td>
+                <td>function</td>
+                <td>
+                  Callback function to get the updated query object. This prop is required to get
+                  the updated query object.
                 </td>
               </tr>
               <tr>
@@ -373,6 +395,13 @@ const App: Component = () => {
                 <td>
                   Show the not toggle for rules and groups. This prop is optional. Default is{' '}
                   <code>both</code>.
+                </td>
+              </tr>
+              <tr>
+                <td>showBranches</td>
+                <td>boolean</td>
+                <td>
+                  Show branches for groups. This prop is optional. Default is <code>true</code>.
                 </td>
               </tr>
             </tbody>