+++ src/hooks/useReducer.ts
@@ -5,6 +5,7 @@ type Reducer<State, Action> = (state: State, action: Action) => State;
 export const useReducer = <State extends object, Action>(
   reducer: Reducer<State, Action>,
   initialState: State,
+  triggerUpdateCallBack: () => void,
 ): [State, (action: Action) => void] => {
   const [store, setStore] = createStore<State>(initialState);
 
@@ -14,6 +15,8 @@ export const useReducer = <State extends object, Action>(
         reducer(store, action);
       }),
     );
+    // workaround to trigger update related to store...
+    triggerUpdateCallBack();
   };
 
   return [store, dispatch];