+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Initial major release for Query Builder.
+
 ## 1.0.0-beta.0 - 2024-11-12
+
+- Initial beta release
 
 ## 0.0.1-beta.1 - 2024-11-10
 - Initial beta release of Query Builder component in SolidJS.